renaming component to be same as before
