+++ Assets/Mirage/Runtime/Logging/LogSettings.cs
@@ -1,101 +1,46 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using UnityEngine;
 
 namespace Mirage.Logging
 {
-    public class LogSettings : ScriptableObject
+    /// <summary>
+    /// Used to load LogSettings in build
+    /// </summary>
+    [DisallowMultipleComponent]
+    [AddComponentMenu("Network/NetworkLogSettings")]
+    [HelpURL("https://mirror-networking.com/docs/Articles/Components/NetworkLogSettings.html")]
+    public class LogSettings : MonoBehaviour
     {
-        public List<LoggerSettings> LogLevels = new List<LoggerSettings>();
-
-        [Serializable]
-        public class LoggerSettings
-        {
-            public string Name;
-            public string Namespace;
-            public string FullName => $"{Namespace}.{Name}";
-
-            public LogType logLevel;
-
-            public LoggerSettings(string name, string Namespace, LogType level)
-            {
-                Name = name;
-                this.Namespace = Namespace;
-                logLevel = level;
-            }
-            public LoggerSettings(string fullname, LogType level)
-            {
-                (Name, Namespace) = GetNameAndNameSapceFromFullname(fullname);
-                logLevel = level;
-            }
-
-            private static (string name, string @namespace) GetNameAndNameSapceFromFullname(string fullname)
-            {
-                string[] parts = fullname.Split('.');
-                string name = parts.Last();
-
-                string @namespace;
-                if (parts.Length == 1)
-                {
-                    @namespace = string.Empty;
-                }
-                else
-                {
-                    @namespace = string.Join(".", parts.Take(parts.Length - 1));
-                }
-
-                return (name, @namespace);
-            }
-        }
-    }
-
-    public static class LogSettingsExtensions
-    {
-        public static void SaveFromDictionary(this LogSettings settings, Dictionary<string, ILogger> dictionary)
-        {
-            if (settings == null)
-            {
-                Debug.LogWarning("Could not SaveFromDictionary because LogSettings were null");
-                return;
-            }
-
-            settings.LogLevels.Clear();
-
-            foreach (KeyValuePair<string, ILogger> kvp in dictionary)
-            {
-                settings.LogLevels.Add(new LogSettings.LoggerSettings(kvp.Key, kvp.Value.filterLogType));
-            }
+        [Header("Log Settings Asset")]
+        [SerializeField] internal LogSettingsSO settings;
 
 #if UNITY_EDITOR
-            UnityEditor.EditorUtility.SetDirty(settings);
+        // called when component is added to GameObject
+        void Reset()
+        {
+            LogSettingsSO existingSettings = EditorLogSettingsLoader.FindLogSettings();
+            if (existingSettings != null)
+            {
+                settings = existingSettings;
+
+                UnityEditor.EditorUtility.SetDirty(this);
+            }
+        }
 #endif
+
+        void Awake()
+        {
+            RefreshDictionary();
         }
 
-        public static void LoadIntoDictionary(this LogSettings settings, Dictionary<string, ILogger> dictionary)
+        void OnValidate()
         {
-            if (settings == null)
-            {
-                Debug.LogWarning("Could not LoadIntoDictionary because LogSettings were null");
-                return;
-            }
+            // if settings field is changed
+            RefreshDictionary();
+        }
 
-            foreach (LogSettings.LoggerSettings logLevel in settings.LogLevels)
-            {
-                if (dictionary.TryGetValue(logLevel.FullName, out ILogger logger))
-                {
-                    logger.filterLogType = logLevel.logLevel;
-                }
-                else
-                {
-                    logger = new Logger(Debug.unityLogger)
-                    {
-                        filterLogType = logLevel.logLevel
-                    };
-
-                    dictionary[logLevel.FullName] = logger;
-                }
-            }
+        void RefreshDictionary()
+        {
+            settings.LoadIntoDictionary(LogFactory.loggers);
         }
     }
 }
