+++ Assets/Mirage/Runtime/Logging/LogSettings.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ac6e8eccf4b6f4dc7b24c276ef47fde8
+guid: 8af93c1717d353e49ba77dae14e27387
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
